feat(order): add order type to retail orders
This commit adds the `order.setType(GlobalConfig.ORDER_TYPE_RETAIL)` line to the `CartController` to set the order type for retail orders. Additionally, the SQL query in `OrderDAO` is updated to include the `type` field when inserting a new order into the database. These changes ensure that the order type is correctly stored and retrieved for retail orders.